'gtk2_ardour' - Add namespaces + casting where necessary + general bits of 'correctness' to keep MSVC happy

+++ gtk2_ardour/audio_clock.cc
@@ -1128,7 +1128,7 @@ AudioClock::set_minsec (framepos_t when, bool /*force*/)
 	mins = (int) floor (left / (_session->frame_rate() * 60.0f));
 	left -= (framecnt_t) floor (mins * _session->frame_rate() * 60.0f);
 	secs = (int) floor (left / (float) _session->frame_rate());
-	left -= (framecnt_t) floor (secs * _session->frame_rate());
+	left -= (framecnt_t) floor ((double)(secs * _session->frame_rate()));
 	millisecs = floor (left * 1000.0 / (float) _session->frame_rate());
 
 	if (negative) {
